+++ lib/active_attr/attributes.rb
@@ -120,14 +120,6 @@ module ActiveAttr
     end
     alias_method :[]=, :write_attribute
 
-    protected
-
-    # Overrides ActiveModel::AttributeMethods
-    # @private
-    def attribute_method?(attr_name)
-      self.class.attribute_names.include? attr_name.to_s
-    end
-
     private
 
     # Read an attribute from the attributes hash
@@ -150,8 +142,9 @@ module ActiveAttr
       # Defines an attribute
       #
       # For each attribute that is defined, a getter and setter will be
-      # added as an instance method to the model. An AttributeDefinition
-      # instance will be added to result of the attributes class method.
+      # added as an instance method to the model. An
+      # {AttributeDefinition} instance will be added to result of the
+      # attributes class method.
       #
       # @example Define an attribute.
       #   attribute :name
@@ -161,8 +154,34 @@ module ActiveAttr
       # @raise [DangerousAttributeError] if the attribute name conflicts with
       #   existing methods
       #
+      # @return [AttributeDefinition] Attribute's definition
+      #
       # @since 0.2.0
       def attribute(name, options={})
+        if dangerous_attribute_method_name = dangerous_attribute?(name)
+          raise DangerousAttributeError, %{an attribute method named "#{dangerous_attribute_method_name}" would conflict with an existing method}
+        else
+          attribute! name, options
+        end
+      end
+
+      # Defines an attribute without checking for conflicts
+      #
+      # Allows you to define an attribute whose methods will conflict
+      # with an existing method. For example, Ruby's Timeout library
+      # adds a timeout method to Object. Attempting to define a timeout
+      # attribute using .attribute will raise a
+      # {DangerousAttributeError}, but .attribute! will not.
+      #
+      # @example Define a dangerous attribute.
+      #   attribute! :timeout
+      #
+      # @param (see AttributeDefinition#initialize)
+      #
+      # @return [AttributeDefinition] Attribute's definition
+      #
+      # @since 0.6.0
+      def attribute!(name, options={})
         AttributeDefinition.new(name, options).tap do |attribute_definition|
           attribute_name = attribute_definition.name.to_s
           # Force active model to generate attribute methods
@@ -197,6 +216,30 @@ module ActiveAttr
         @attributes ||= ActiveSupport::HashWithIndifferentAccess.new
       end
 
+      # Determine if a given attribute name is dangerous
+      #
+      # Some attribute names can cause conflicts with existing methods
+      # on an object. For example, an attribute named "timeout" would
+      # conflict with the timeout method that Ruby's Timeout library
+      # mixes into Object.
+      #
+      # @example Testing a harmless attribute
+      #   Person.dangerous_attribute? :name #=> false
+      #
+      # @example Testing a dangerous attribute
+      #   Person.dangerous_attribute? :nil #=> "nil?"
+      #
+      # @param name Attribute name
+      #
+      # @return [false, String] False or the conflicting method name
+      #
+      # @since 0.6.0
+      def dangerous_attribute?(name)
+        attribute_methods(name).detect do |method_name|
+          !DEPRECATED_OBJECT_METHODS.include?(method_name.to_s) && allocate.respond_to?(method_name, true)
+        end unless attribute_names.include? name.to_s
+      end
+
       # Returns the class name plus its attribute names
       #
       # @example Inspect the model's definition.
@@ -208,7 +251,7 @@ module ActiveAttr
       def inspect
         inspected_attributes = attribute_names.sort
         attributes_list = "(#{inspected_attributes.join(", ")})" unless inspected_attributes.empty?
-        "#{self.name}#{attributes_list}"
+        "#{name}#{attributes_list}"
       end
 
       protected
@@ -223,16 +266,14 @@ module ActiveAttr
         @attributes = attributes
       end
 
-      # Overrides ActiveModel::AttributeMethods
-      # @private
-      def instance_method_already_implemented?(method_name)
-        deprecated_object_method = DEPRECATED_OBJECT_METHODS.include?(method_name.to_s)
-        already_implemented = !deprecated_object_method && self.allocate.respond_to?(method_name, true)
-        raise DangerousAttributeError, %{an attribute method named "#{method_name}" would conflict with an existing method} if already_implemented
-        false
-      end
-
       private
+
+      # Expand an attribute name into its generated methods names
+      #
+      # @since 0.6.0
+      def attribute_methods(name)
+        attribute_method_matchers.map { |matcher| matcher.method_name name }
+      end
 
       # Ruby inherited hook to assign superclass attributes to subclasses
       #